pyproject.toml
[project]
name = "diffusion-language-model"
version = "0.1.0"
description = "MDLM (Masked Diffusion Language Model) pre-training"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.10.8",
]

[project.scripts]
train = "src.train:main"
sample = "src.sample:main"
